ref: fix types for sentry.integrations.utils.sync
