+++ pyproject.toml
@@ -123,7 +123,6 @@ module = [
     "sentry.integrations.pagerduty.actions.form",
     "sentry.integrations.slack.message_builder.notifications.issues",
     "sentry.integrations.slack.webhooks.event",
-    "sentry.integrations.utils.sync",
     "sentry.issues.search",
     "sentry.middleware.auth",
     "sentry.middleware.ratelimit",
@@ -251,6 +250,7 @@ module = [
     "sentry.integrations.slack.threads.*",
     "sentry.integrations.slack.views.*",
     "sentry.integrations.source_code_management.repository",
+    "sentry.integrations.utils.sync",
     "sentry.integrations.vsts.actions.*",
     "sentry.integrations.vsts.tasks.*",
     "sentry.issues",
