+++ src/sentry/integrations/mixins/issues.py
@@ -7,30 +7,35 @@ from collections import defaultdict
 from collections.abc import Mapping, Sequence
 from copy import deepcopy
 from operator import attrgetter
-from typing import Any, ClassVar
+from typing import TYPE_CHECKING, Any, ClassVar
 
+from sentry import features
 from sentry.eventstore.models import GroupEvent
 from sentry.integrations.base import IntegrationInstallation
 from sentry.integrations.models.external_issue import ExternalIssue
+from sentry.integrations.models.integration import Integration
 from sentry.integrations.services.assignment_source import AssignmentSource
 from sentry.integrations.services.integration import integration_service
 from sentry.integrations.tasks.sync_status_inbound import (
     sync_status_inbound as sync_status_inbound_task,
 )
-from sentry.integrations.utils.sync import where_should_sync
 from sentry.issues.grouptype import GroupCategory
 from sentry.issues.issue_occurrence import IssueOccurrence
 from sentry.models.group import Group
 from sentry.models.grouplink import GroupLink
+from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.notifications.utils import get_notification_group_title
 from sentry.shared_integrations.exceptions import IntegrationError
-from sentry.silo.base import all_silo_function
+from sentry.silo.base import all_silo_function, region_silo_function
 from sentry.users.models.user import User
 from sentry.users.services.user import RpcUser
 from sentry.users.services.user_option import get_option_from_list, user_option_service
 from sentry.utils.http import absolute_uri
 from sentry.utils.safe import safe_execute
+
+if TYPE_CHECKING:
+    from sentry.integrations.services.integration import RpcIntegration
 
 logger = logging.getLogger("sentry.integrations.issues")
 MAX_CHAR = 50
@@ -369,6 +374,36 @@ class IssueBasicIntegration(IntegrationInstallation, ABC):
 
     def update_comment(self, issue_id, user_id, group_note):
         pass
+
+
+@region_silo_function
+def where_should_sync(
+    integration: RpcIntegration | Integration,
+    key: str,
+    organization_id: int | None = None,
+) -> Sequence[Organization]:
+    """
+    Given an integration, get the list of organizations where the sync type in
+    `key` is enabled. If an optional `organization_id` is passed, then only
+    check the integration for that organization.
+    """
+    kwargs = dict()
+    if organization_id is not None:
+        kwargs["id"] = organization_id
+        ois = integration_service.get_organization_integrations(
+            integration_id=integration.id, organization_id=organization_id
+        )
+    else:
+        ois = integration_service.get_organization_integrations(integration_id=integration.id)
+
+    organizations = Organization.objects.filter(id__in=[oi.organization_id for oi in ois])
+    ret = []
+    for organization in organizations.filter(**kwargs):
+        if features.has("organizations:integrations-issue-sync", organization):
+            installation = integration.get_installation(organization_id=organization.id)
+            if isinstance(installation, IssueBasicIntegration) and installation.should_sync(key):
+                ret.append(organization)
+    return ret
 
 
 class IntegrationSyncTargetNotFound(IntegrationError):
